// src/types/utils.ts
export interface CatalogDTO {
	id?: string;
	group?: string;
	key?: string;
	position?: string;
	dataIndex?: string;
	brand?: string;
	attribute?: string;
	name?: string;
	internationalCode?: string;
	fixUnits?: any;
	units?: any;
	recommendedUnits?: any;
	qrCode?: any;
	shtrixCode?: any;
    title?: string;
}